improved memory estimate code and print style
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@3587 6419b465-d902-0410-997c-be2266baf723

+++ gpaw/grid_descriptor.py
@@ -512,7 +512,7 @@ class GridDescriptor(Domain):
 
     def bytecount(self, dtype=float):
         """Get the number of bytes used by a grid of specified dtype."""
-        return np.prod(self.n_c) * np.array(1, dtype).itemsize
+        return long(np.prod(self.n_c)) * np.array(1, dtype).itemsize
 
 
 class RadialGridDescriptor:
